=More SPI cleanup

+++ stm32f4/src/spi.h
@@ -129,7 +129,7 @@ void spiWait(const extDevice_t *dev);
 
 void spiSequence(const extDevice_t *dev, busSegment_t *segments);
 
-bool spiSetBusInstance(extDevice_t *dev, const uint32_t device);
+bool spiSetBusInstance(extDevice_t *dev, const uint8_t device);
 
 void spiSetClkDivisor(const extDevice_t *dev, const uint16_t divider);
